// src/generated/models/knowledge-article-model.ts
import type { KnowledgeCategory } from './knowledge-category-model';
import type { Person } from './person-model';
import type { Workspace } from './workspace-model';

export const KnowledgeArticleAudienceKeyToLabel = {
  'Requesters': 'Requesters',
  'Agents': 'Agents',
  'AllMembers': 'All Members'
} as const;
export type KnowledgeArticleAudienceKey = keyof typeof KnowledgeArticleAudienceKeyToLabel;

export const KnowledgeArticleStateKeyToLabel = {
  'Draft': 'Draft',
  'Review': 'Review',
  'Published': 'Published',
  'Archived': 'Archived'
} as const;
export type KnowledgeArticleStateKey = keyof typeof KnowledgeArticleStateKeyToLabel;

export interface KnowledgeArticle {
  /**
   * @displayName Knowledge Article ID
   * @description Unique system identifier - GUID generated by system
   * @uiConstraints NEVER create input fields for this ID
   */
  id: string;
  /**
   * @displayName Article Number
   * @description Human-readable record identifier
   * @validationRule Required for create/update operations
   */
  articleNumber: string;
  /**
   * @displayName Audience
   * @validationRule Required for create/update operations
   */
  audienceKey: KnowledgeArticleAudienceKey;
  /**
   * @displayName Created and Updated Audit Fields
   * @validationRule Required for create/update operations
   */
  createdAndUpdatedAuditFields: string;
  /**
   * @displayName Current Version Number
   * @validationRule Required for create/update operations
   */
  currentVersionNumber: number;
  /**
   * @displayName Knowledge Category
   * @validationRule Required for create/update operations
   */
  knowledgeCategory: Pick<KnowledgeCategory, 'id' | 'name1'>;
  /**
   * @displayName Owner Person
   * @validationRule Required for create/update operations
   */
  ownerPerson: Pick<Person, 'id' | 'displayName'>;
  /**
   * @displayName Published and Review Dates
   * @validationRule Required for create/update operations
   */
  publishedAndReviewDates: string;
  /**
   * @displayName State
   * @validationRule Required for create/update operations
   */
  stateKey: KnowledgeArticleStateKey;
  /**
   * @displayName Summary
   * @validationRule Required for create/update operations
   */
  summary: string;
  /**
   * @displayName Tags
   * @validationRule Required for create/update operations
   */
  tags: string;
  /**
   * @displayName Title
   * @validationRule Required for create/update operations
   */
  title: string;
  /**
   * @displayName View and Feedback Counts
   * @validationRule Required for create/update operations
   */
  viewAndFeedbackCounts: string;
  /**
   * @displayName Workspace
   * @validationRule Required for create/update operations
   */
  workspace: Pick<Workspace, 'id' | 'workspaceName'>;
}

export const _KnowledgeArticle = 'KnowledgeArticle' as const;