// apps/enterprise-service-management/src/generated/models/knowledge-category-model.ts
import type { Workspace } from './workspace-model';

export interface KnowledgeCategory {
  /**
   * @displayName Knowledge Category ID
   * @description Unique system identifier - GUID generated by system
   * @uiConstraints NEVER create input fields for this ID
   */
  id: string;
  /**
   * @displayName Name
   * @description Human-readable record identifier
   * @validationRule Required for create/update operations
   */
  name1: string;
  /**
   * @displayName Active
   * @validationRule Required for create/update operations
   */
  active: boolean;
  /**
   * @displayName Description
   * @validationRule Required for create/update operations
   */
  description: string;
  /**
   * @displayName Icon
   * @validationRule Required for create/update operations
   */
  icon: string;
  /**
   * @displayName Parent Category
   */
  parentCategory?: Pick<KnowledgeCategory, 'id' | 'name1'>;
  /**
   * @displayName Sort Order
   * @validationRule Required for create/update operations
   */
  sortOrder: number;
  /**
   * @displayName Workspace
   * @validationRule Required for create/update operations
   */
  workspace: Pick<Workspace, 'id' | 'workspaceName'>;
}

export const _KnowledgeCategory = 'KnowledgeCategory' as const;