// src/generated/models/knowledge-version-model.ts
import type { Person } from './person-model';
import type { KnowledgeArticle } from './knowledge-article-model';
import type { Workspace } from './workspace-model';

export const KnowledgeVersionStateKeyToLabel = {
  'Draft': 'Draft',
  'Review': 'Review',
  'Published': 'Published',
  'Archived': 'Archived'
} as const;
export type KnowledgeVersionStateKey = keyof typeof KnowledgeVersionStateKeyToLabel;

export interface KnowledgeVersion {
  /**
   * @displayName Knowledge Version ID
   * @description Unique system identifier - GUID generated by system
   * @uiConstraints NEVER create input fields for this ID
   */
  id: string;
  /**
   * @displayName Immutable Snapshot Label
   * @description Human-readable record identifier
   * @validationRule Required for create/update operations
   */
  immutableSnapshotLabel: string;
  /**
   * @displayName Author
   * @validationRule Required for create/update operations
   */
  author: Pick<Person, 'id' | 'displayName'>;
  /**
   * @displayName Body Content
   * @validationRule Required for create/update operations
   */
  bodyContent: string;
  /**
   * @displayName Change Notes
   * @validationRule Required for create/update operations
   */
  changeNotes: string;
  /**
   * @displayName Knowledge Article
   * @validationRule Required for create/update operations
   */
  knowledgeArticle: Pick<KnowledgeArticle, 'id' | 'articleNumber'>;
  /**
   * @displayName Review Comments
   * @validationRule Required for create/update operations
   */
  reviewComments: string;
  /**
   * @displayName Reviewer
   * @validationRule Required for create/update operations
   */
  reviewer: Pick<Person, 'id' | 'displayName'>;
  /**
   * @displayName State
   * @validationRule Required for create/update operations
   */
  stateKey: KnowledgeVersionStateKey;
  /**
   * @displayName Submitted, Reviewed, and Published Timestamps
   * @validationRule Required for create/update operations
   */
  submittedReviewedAndPublishedTimestamps: string;
  /**
   * @displayName Summary
   * @validationRule Required for create/update operations
   */
  summary: string;
  /**
   * @displayName Title
   * @validationRule Required for create/update operations
   */
  title: string;
  /**
   * @displayName Version Number
   * @validationRule Required for create/update operations
   */
  versionNumber: number;
  /**
   * @displayName Workspace
   * @validationRule Required for create/update operations
   */
  workspace: Pick<Workspace, 'id' | 'workspaceName'>;
}

export const _KnowledgeVersion = 'KnowledgeVersion' as const;